more options and wavefront support
fixed the wavefront generation so now it works again. had to tweak the offset and I made it work using greedy meshing (file size efficiency reasons)

also made the size of the chunkarray customizable, but it HAS to be a power of 2 and ros2 crashes the program if the value is too big.

also added a parameter to set the filepath the wavefront should be written to.